feat: add support for CoseEncrypt0 using backend
